Move pytest fixtures around for new tests
