+++ tests/test_internal_plugins.py
@@ -1,41 +1,13 @@
-import pytest
+# import pytest
 from carbon_txt import validators  # type: ignore
 import logging
 
 logger = logging.getLogger(__name__)
 
 
-@pytest.fixture()
-def reset_plugin_registry():
-    """
-    Reset the plugin registry to the default state. We need to do this
-    because the plugin framework we have doesn't reset on each test, and
-    if we set up a plugin in one test, it can still be active in the next.
-    """
-    from carbon_txt.plugins import pm
-
-    modules = pm.get_plugins()
-    print(f"\n current modules: {modules}")
-    for mod in modules:
-        print(f"\n Unregistering plugin {mod}")
-        pm.unregister(mod)
-
-    print(f"\n updated: {pm.get_plugins()}")
-
-    return pm
-
-
-@pytest.fixture()
-def settings_with_active_csrd_plugin(reset_plugin_registry, settings):
-    settings.ACTIVE_CARBON_TXT_PLUGINS = ["carbon_txt.process_csrd_document"]
-
-    return settings
-
-
 class TestCarbonTxtValidatorWithCSRDPlugin:
     def test_validate_carbon_txt_file_with_csrd_report(
-        self,
-        settings_with_active_csrd_plugin,
+        self, settings_with_active_csrd_greenweb_plugin, settings
     ):
         """
         Test that we can validate a carbon.txt file valid CSRD report, and that
@@ -43,7 +15,7 @@ class TestCarbonTxtValidatorWithCSRDPlugin:
         """
 
         validator = validators.CarbonTxtValidator(
-            active_plugins=settings_with_active_csrd_plugin.ACTIVE_CARBON_TXT_PLUGINS
+            active_plugins=settings.ACTIVE_CARBON_TXT_PLUGINS
         )
 
         test_domain = "used-in-tests.carbontxt.org"
